Create buildCommands.
